+++ packages/core-js/modules/es.string.replace.js
@@ -76,10 +76,12 @@ fixRegExpWellKnownSymbolLogic('replace', function (_, nativeReplace, maybeCallNa
       var rx = anObject(this);
       var S = toString(string);
 
+      var flags = toString(getRegExpFlags(rx));
       if (
         typeof replaceValue == 'string' &&
         !~stringIndexOf(replaceValue, UNSAFE_SUBSTITUTE) &&
-        !~stringIndexOf(replaceValue, '$<')
+        !~stringIndexOf(replaceValue, '$<') &&
+        !~stringIndexOf(flags, 'y')
       ) {
         var res = maybeCallNative(nativeReplace, rx, S, replaceValue);
         if (res.done) return res.value;
@@ -88,7 +90,6 @@ fixRegExpWellKnownSymbolLogic('replace', function (_, nativeReplace, maybeCallNa
       var functionalReplace = isCallable(replaceValue);
       if (!functionalReplace) replaceValue = toString(replaceValue);
 
-      var flags = toString(getRegExpFlags(rx));
       var global = !!~stringIndexOf(flags, 'g');
       var fullUnicode;
       if (global) {